Replaced long with int (for sizes).

+++ cscharls/ByteStreamInfo.cs
@@ -6,22 +6,22 @@ using System.IO;
 
 namespace CharLS
 {
-    //
+    // 
     // ByteStreamInfo & FromByteArray helper function
-    //
+    // 
     // ByteStreamInfo describes the stream: either set rawStream to a valid stream, or rawData/count, not both.
     // it's possible to decode to memorystreams, but using rawData will always be faster.
-    //
+    // 
     // Example use:
-    //     ByteStreamInfo streamInfo = { fileStream.rdbuf() };
+    // ByteStreamInfo streamInfo = { fileStream.rdbuf() };
     // or
-    //     ByteStreamInfo streamInfo = FromByteArray( bytePtr, byteCount);
-    //
+    // ByteStreamInfo streamInfo = FromByteArray( bytePtr, byteCount);
+    // 
     public class ByteStreamInfo
     {
         private byte[] _rawData;
 
-        private long _position;
+        private int _position;
 
         public ByteStreamInfo(Stream stream)
         {
@@ -30,10 +30,10 @@ namespace CharLS
             rawStream = stream;
             _rawData = null;
             _position = 0;
-            count = stream.CanSeek ? stream.Length : -1;
+            count = stream.CanSeek ? (int)stream.Length : -1;
         }
 
-        public ByteStreamInfo(byte[] bytes, long length = -1)
+        public ByteStreamInfo(byte[] bytes, int length = -1)
         {
             if (bytes == null) throw new ArgumentNullException(nameof(bytes));
 
@@ -45,16 +45,16 @@ namespace CharLS
 
         public Stream rawStream { get; }
 
-        public ArraySegment<byte> rawData => new ArraySegment<byte>(_rawData, (int)_position, (int)count);
+        public ArraySegment<byte> rawData => new ArraySegment<byte>(_rawData, _position, count);
 
-        public long count { get; private set; }
+        public int count { get; private set; }
 
-        public static ByteStreamInfo FromByteArray(byte[] bytes, long count)
+        public static ByteStreamInfo FromByteArray(byte[] bytes, int count)
         {
             return new ByteStreamInfo(bytes, count);
         }
 
-        public void Seek(long skip)
+        public void Seek(int skip)
         {
             if (_rawData == null) return;
