+++ cscharls/util.cs
@@ -67,7 +67,7 @@ namespace CharLS
             return i;
         }
 
-        public static void SkipBytes(ByteStreamInfo streamInfo, long count)
+        public static void SkipBytes(ByteStreamInfo streamInfo, int count)
         {
             streamInfo.Seek(count);
         }
